filtering metrics based on subexpression dimenssions and
listening to events topic alamr-definition-* events

However this approach is not usable as tables cannot be create
dynamicaly, so there cannot be a table for each metric needed by
subexpression of alarm created dynamically.

So there has to be one global table which can be sharded by
subexpression id. That is the next trial implementation.